fixup! [AIRFLOW-5894] Group tests for the Dags command
